feat(amount.js): add amount upgrade functionality with increment and update methods to track upgrades refactor(buyUpgrade.js): remove buyUpgrade.js and integrate its logic into upgrade.js for better organization refactor(card.js): update createUpgradeCard to return amount along with the upgrade card for better data handling refactor(content.js): modify createUpgradeContent to return an object containing content and amount for improved structure refactor(upgrade.js): enhance buyUpgrade function to include amount tracking and update logic for better upgrade management
